commits second and third resubmitt
I am still confused as to what exactly you are wanting. The first example that you gave us that you wanted us to complete was just a text file without any CSS attached, and the second file is a standard picture of a box-frame model. I am confused as to what this lesson is supposed to teach me? I am still very perplexed as to why my first assignment was rejected. All of the elements were there I just elaborated and added CSS. It seems as if there is an embedded preference in the file that you sent us that doesn't allow for you to use your own indention in the file. it pretty much does it for you, well at least for a simple box model, otherwise, it doesn't work.  Maybe in class, you can explain why we are using this teaching method because I am unsure what I am supposed to learn from this lesson.